+++ tests/test_cp.cpp
@@ -25,9 +25,9 @@ int main(int argc, char **argv) {
     Vector<double> x(3 * m.n);
     double         r = 1.0 / sqrt(m.n);
 
-    for (unsigned i = 0; i < m.n; ++i) {
-        x[3 * i]     = m.v[i]->z.real() / (1 - r);
-        x[3 * i + 1] = m.v[i]->z.imag() / (1 - r);
+    for (int64_t i = 0; i < int(m.n); ++i) {
+        x[3 * i]     = m.v[size_t(i)]->z.real() / (1 - r);
+        x[3 * i + 1] = m.v[size_t(i)]->z.imag() / (1 - r);
         x[3 * i + 2] = .8 * r;
     }
 
@@ -43,9 +43,9 @@ int main(int argc, char **argv) {
     spdlog::info("Final value of f:      {}", MM.fx);
     spdlog::info("Final square gradient: {}", MM.gx.squaredNorm());
 
-    for (unsigned i = 0; i < m.n; ++i) {
-        m.v[i]->z = cpx(x[3 * i], x[3 * i + 1]);
-        m.v[i]->r = x[3 * i + 2];
+    for (size_t i = 0; i < m.n; ++i) {
+        m.v[i]->z = cpx(x[3 * int(i)], x[3 * int(i) + 1]);
+        m.v[i]->r = x[3 * int(i) + 2];
     }
 
     Figure f{H.title};
